chore: Rename global-fetch to global-node-fetch
