+++ packages/@pollyjs/adapter-node-http/tests/integration/adapter-node-fetch-test.js
@@ -1,4 +1,4 @@
-import '@pollyjs-tests/helpers/global-fetch';
+import '@pollyjs-tests/helpers/global-node-fetch';
 
 import setupFetchRecord from '@pollyjs-tests/helpers/setup-fetch-record';
 import adapterTests from '@pollyjs-tests/integration/adapter-tests';
